refactor(config): move deployment loading to deployment module
